Pull CAPTCHAs into their own control
Show audio if in full trust
Allow localizing image alt text
