+++ Controls/App_LocalResources/Captchas.ascx.resx
@@ -117,43 +117,25 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="btnEmailAFriend.Text" xml:space="preserve">
-    <value>Email A Friend</value>
+  <data name="InvisibleCaptchaFailed.Text" xml:space="preserve">
+    <value>A hidden text box was filled in, indicating that this form was not filled in by hand.  To prevent SPAM being sent through this form, we have to disallow any submissions that aren't entered by hand.</value>
   </data>
-  <data name="lblTo.Text" xml:space="preserve">
-    <value>Enter e-mail address of recipient:</value>
+  <data name="InvisibleCaptchaLabel.Text" xml:space="preserve">
+    <value>Do not fill this textbox.</value>
   </data>
-  <data name="lblFrom.Text" xml:space="preserve">
-    <value>Enter your e-mail address:</value>
+  <data name="TimeoutCaptchaFailed.Text" xml:space="preserve">
+    <value>To prevent SPAM, you must wait before submitting the form.</value>
   </data>
-  <data name="lblMessage.Text" xml:space="preserve">
-    <value>Message to attach:</value>
+  <data name="StandardCaptchaFailed.Text" xml:space="preserve">
+    <value>Sorry, your entry didn't match the code; please try again</value>
   </data>
-  <data name="btnSend.Text" xml:space="preserve">
-    <value>Send Email</value>
+  <data name="StandardCaptchaLabel.Text" xml:space="preserve">
+    <value>Type the code from the image</value>
   </data>
-  <data name="btnCancel.Text" xml:space="preserve">
-    <value>Cancel</value>
+  <data name="StandardCaptchaLink.Text" xml:space="preserve">
+    <value>Generate New Image</value>
   </data>
-  <data name="lblPrivacy.Text" xml:space="preserve">
-    <value>We use this address only to tell the recipient who sent the message. We do not save or re-use it in any way.</value>
-  </data>
-  <data name="lblMultiple.Text" xml:space="preserve">
-    <value>(Use a comma to separate multiple e-mail addresses.)</value>
-  </data>
-  <data name="EmailAFriend.Text" xml:space="preserve">
-    <value>&lt;html&gt;&lt;body&gt;Dear [Engage:Recipient]&lt;br&gt;&lt;a href=[Engage:Url]&gt; This link &lt;/a&gt; has been sent to you from [Engage:From] with the attached message. &lt;br /&gt;&lt;br /&gt;[Engage:Message]&lt;/body&gt;&lt;/html&gt;</value>
-  </data>
-  <data name="EmailAFriendSubject.Text" xml:space="preserve">
-    <value>[Engage:Portal] - A Friend has sent you a link to an Article</value>
-  </data>
-  <data name="Please enter one or more recipients.Text" xml:space="preserve">
-    <value>Please enter one or more recipients</value>
-  </data>
-  <data name="Please enter your email address.Text" xml:space="preserve">
-    <value>Please enter your email address</value>
-  </data>
-  <data name="ValidationSummary.Header" xml:space="preserve">
-    <value>&lt;strong&gt;Please correct the following error(s):&lt;/strong&gt;</value>
+  <data name="StandardCaptchaImageAltText.Text" xml:space="preserve">
+    <value />
   </data>
 </root>
