+++ Controls/EmailAFriend.ascx.cs
@@ -17,12 +17,9 @@ namespace Engage.Dnn.Publish.Controls
     using System.Text;
     using System.Web.UI;
 
-    using DotNetNuke.Framework;
     using DotNetNuke.Services.Exceptions;
     using DotNetNuke.Services.Localization;
     using DotNetNuke.Services.Mail;
-
-    using Engage.Dnn.Publish.Util;
 
     public partial class EmailAFriend : ModuleBase
     {
@@ -89,8 +86,7 @@ namespace Engage.Dnn.Publish.Controls
                 this.AddJQueryReference();
                 this.Page.ClientScript.RegisterClientScriptInclude("Engage_Publish_ModalPopup", this.ResolveUrl("../Scripts/ModalPopup.js"));
                 this.EmailAFriendPopupTriggerLink.Attributes["data-modal-target-id"] = this.pnlEmailAFriend.ClientID;
-                this.LocalizeValidators();
-                this.EnableValidators();
+                this.ValidationSummary.HeaderText = Localization.GetString("ValidationSummary.Header", this.LocalResourceFile);
                 this.SetValidationGroup();
 
                 if (!this.IsPostBack)
@@ -104,32 +100,12 @@ namespace Engage.Dnn.Publish.Controls
             }
         }
 
-        private void LocalizeValidators()
-        {
-            this.ValidationSummary.HeaderText = Localization.GetString("ValidationSummary.Header", this.LocalResourceFile);
-            this.InvisibleCaptcha.ErrorMessage = Localization.GetString("InvisibleCaptchaFailed", this.LocalResourceFile);
-            this.InvisibleCaptcha.InvisibleTextBoxLabel = Localization.GetString("InvisibleCaptchaLabel", this.LocalResourceFile);
-            this.TimeoutCaptcha.ErrorMessage = Localization.GetString("TimeoutCaptchaFailed", this.LocalResourceFile);
-            this.StandardCaptcha.ErrorMessage = Localization.GetString("StandardCaptchaFailed", this.LocalResourceFile);
-            this.StandardCaptcha.CaptchaLinkButtonText = Localization.GetString("StandardCaptchaLink", this.LocalResourceFile);
-            this.StandardCaptcha.CaptchaTextBoxLabel = Localization.GetString("StandardCaptchaLabel", this.LocalResourceFile);
-        }
-
-        private void EnableValidators()
-        {
-            this.InvisibleCaptcha.Visible = this.InvisibleCaptcha.Enabled = Utility.GetBooleanPortalSetting(Utility.PublishEnableInvisibleCaptcha, this.PortalId, true);
-            this.TimeoutCaptcha.Visible = this.TimeoutCaptcha.Enabled = Utility.GetBooleanPortalSetting(Utility.PublishEnableTimedCaptcha, this.PortalId, true);
-            this.StandardCaptcha.Visible = this.StandardCaptcha.Enabled = Utility.GetBooleanPortalSetting(Utility.PublishEnableStandardCaptcha, this.PortalId, false);
-        }
-
         private void SetValidationGroup()
         {
             var validationGroup = "ValidateGroupSend" + this.ModuleId.ToString(CultureInfo.InvariantCulture);
             this.ToRequiredValidator.ValidationGroup = validationGroup;
             this.FromRequiredValidator.ValidationGroup = validationGroup;
-            this.InvisibleCaptcha.ValidationGroup = validationGroup;
-            this.TimeoutCaptcha.ValidationGroup = validationGroup;
-            this.StandardCaptcha.ValidationGroup = validationGroup;
+            this.Captchas.ValidationGroup = validationGroup;
             this.ValidationSummary.ValidationGroup = validationGroup;
             this.SendButton.ValidationGroup = validationGroup;
         }
